+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/mlapps/mlr/MLRETDataParser.java
@@ -68,10 +68,10 @@ final class MLRETDataParser implements DataParser<MLRData> {
       }
 
       final int[] indices = new int[split.length / 2];
-      final double[] data = new double[split.length / 2];
+      final float[] data = new float[split.length / 2];
       for (int index = 0; index < split.length / 2; ++index) {
         indices[index] = Integer.parseInt(split[2 * index + 1]);
-        data[index] = Double.parseDouble(split[2 * index + 2]);
+        data[index] = Float.parseFloat(split[2 * index + 2]);
       }
 
       final Vector feature = vectorFactory.createSparse(indices, data, numFeatures);
